Fixed directory and syntax errors.
