+++ zoo/animal/Animal.java
@@ -1,0 +1,25 @@
+package zoo.animal;
+
+public abstract class Animal {
+
+    private String location;
+    private String sex;
+
+    public String getLocation() {
+        return location;
+    }
+
+    public void setLocation(String location) {
+        this.location = location;
+    }
+
+    public String getSex() {
+        return sex;
+    }
+
+    public void setSex(String sex) {
+        this.sex = sex;
+    }
+
+    public abstract void makeSound();
+}
