+++ zoo/animal/Mammal.java
@@ -1,0 +1,14 @@
+package zoo.animal;
+
+public abstract class Mammal extends Animal {
+
+    private int numberLegs;
+
+    public int getNumberLegs() {
+        return numberLegs;
+    }
+
+    public void setNumberLegs(int numberLegs) {
+        this.numberLegs = numberLegs;
+    }
+}
