+++ zoo/animal/Reptile.java
@@ -1,0 +1,14 @@
+package zoo.animal;
+
+public abstract class Reptile extends Animal {
+
+    private boolean isVenomous;
+
+    public boolean isVenomous() {
+        return isVenomous;
+    }
+
+    public void setVenomous(boolean isVenomous) {
+        this.isVenomous = isVenomous;
+    }
+}
