+++ zoo/person/Person.java
@@ -1,0 +1,23 @@
+package zoo.person;
+
+public class Person {
+
+    private String name;
+    private String location;
+
+    public String getName() {
+        return this.name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getLocation() {
+        return this.location;
+    }
+
+    public void setLocation(String location) {
+        this.location = location;
+    }
+}
