+++ zoo/person/Visitor.java
@@ -1,0 +1,21 @@
+package zoo.person;
+
+import zoo.ticket.Ticket;
+
+public class Visitor extends Person {
+
+    private Ticket ticket;
+
+    public Ticket getTicket() {
+        return this.ticket;
+    }
+
+    public void setTicket(Ticket ticket) {
+        this.ticket = ticket;
+    }
+
+    @Override
+    public String toString() {
+        return "Visitor name=" + this.getName() + ", location=" + this.getLocation() + ", ticketID=" + ticket.getTicketID();
+    }
+}
